+++ python/agentbay/__init__.py
@@ -22,6 +22,7 @@ from .context_sync import (
     WhiteList,
 )
 from .context_manager import ContextManager, ContextInfoResult, ContextSyncResult
+from .extension import ExtensionsService, ExtensionOption, Extension
 
 __all__ = [
     "Config",
@@ -54,4 +55,7 @@ __all__ = [
     "ContextManager",
     "ContextInfoResult",
     "ContextSyncResult",
+    "ExtensionsService",
+    "ExtensionOption",
+    "Extension",
 ]