+++ python/agentbay/session_params.py
@@ -2,7 +2,7 @@ from typing import Dict, Optional, List, TYPE_CHECKING
 from agentbay.context_sync import ContextSync, SyncPolicy, UploadPolicy, ExtractPolicy, BWList, WhiteList
 
 if TYPE_CHECKING:
-    from agentbay.extention import ExtensionOption
+    from agentbay.extension import ExtensionOption
 
 
 class BrowserContext:
@@ -38,7 +38,7 @@ class BrowserContext:
     Usage Examples:
         ```python
         # With extensions using ExtensionOption
-        from agentbay.extention import ExtensionOption
+        from agentbay.extension import ExtensionOption
         
         ext_option = ExtensionOption(
             context_id="my_extensions",
@@ -89,7 +89,7 @@ class BrowserContext:
         Examples:
             ```python
             # With extensions using ExtensionOption
-            from agentbay.extention import ExtensionOption
+            from agentbay.extension import ExtensionOption
             
             ext_option = ExtensionOption(
                 context_id="my_extensions",